refactor: move the generic path helpers out of the git module
`canonicalize_path` and `home` expand `~/` and `$HOME/` — nothing
about them is git, their callers just happened to be. They join
`tmp_dir` in `src/utils.rs`, and `should_canonicalize` now runs
without the `git` feature too.

That left one function behind, so `git/utils.rs` becomes
`git/clone.rs` — a "utils" file holding a single clone entry point,
next to a crate-level `utils.rs`, was the confusion this cleans up.
